# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: Travelling Llama planner API
  description: Generated from the controllers by OpenApiDocTest. Do not edit by hand.
  version: v1
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/v1/trips/{tripId}/publish-requests:
    get:
      tags:
      - publish-controller
      operationId: requests
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublishView"
    post:
      tags:
      - publish-controller
      operationId: request
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPublishBody"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublishView"
...
  /api/v1/trips/{tripId}/members:
    get:
      tags:
      - trip-controller
      operationId: members
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberView"
    post:
      tags:
      - trip-controller
      operationId: addMember
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberView"
  /api/v1/trips/{tripId}/itinerary:
    get:
      tags:
      - itinerary-controller
      operationId: list_1
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItineraryItem"
    post:
      tags:
      - itinerary-controller
      operationId: create_1
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItineraryItem"
...
  /api/v1/trips/{tripId}/itinerary/{itemId}:
    delete:
      tags:
      - itinerary-controller
      operationId: delete_1
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
    patch:
      tags:
      - itinerary-controller
      operationId: update_1
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ItineraryItem"
...
  /api/v1/trips/{tripId}/flights/lookup:
    get:
      tags:
      - flight-controller
      operationId: lookup
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: query
        required: true
        schema:
          type: string
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LookupResponse"
...
  /api/v1/public/trips/{slug}/flights:
    get:
      tags:
      - public-flight-controller
      operationId: flight
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      - name: number
        in: query
        required: true
        schema:
          type: string
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicFlightView"
...
  /api/v1/trips/{tripId}/itinerary/{itemId}/plan:
    delete:
      tags:
      - itinerary-controller
      operationId: deletePlan
      parameters:
      - name: tripId
        in: path
        required: true
        schema:
          type: string
      - name: itemId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
components:
  schemas:
...
    Airport:
      type: object
      properties:
        iata:
          type: string
        icao:
          type: string
        name:
          type: string
        city:
          type: string
        countryCode:
          type: string
        timezone:
          type: string
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
...
    FlightSnapshot:
      type: object
      properties:
        number:
          type: string
        operatingNumber:
          type: string
        from:
          $ref: "#/components/schemas/Airport"
        to:
          $ref: "#/components/schemas/Airport"
        terminalFrom:
          type: string
        terminalTo:
          type: string
        airline:
          type: string
    ItineraryItem:
      type: object
      properties:
        id:
          type: string
        tripId:
          type: string
        checklistItemId:
          type: string
        planId:
          type: string
        category:
          type: string
          enum:
          - TRANSPORTATION
          - LODGING
          - ACTIVITIES
          - SHOPPING
          - FOOD
          - OTHERS
        description:
          type: string
        note:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        allDay:
          type: boolean
        cost:
          type: number
        currency:
          type: string
        budgetItemId:
          type: string
        countryCodes:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        createdByUserId:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedByUserId:
          type: string
        travellerIds:
          type: array
          items:
            type: string
        flight:
          $ref: "#/components/schemas/FlightSnapshot"
        destinationIds:
          type: array
          writeOnly: true
          items:
            type: string
...
    MemberView:
      type: object
      properties:
        userId:
          type: string
        emailLabel:
          type: string
        displayName:
          type: string
        screenName:
          type: string
        homeCountryCode:
          type: string
        role:
          type: string
          enum:
          - OWNER
          - MEMBER
        hasSignedIn:
          type: boolean
        invitedAt:
          type: string
          format: date-time
...
    PublishRequestView:
      type: object
      properties:
        id:
          type: string
        requestedByUserId:
          type: string
        requestedByDisplayName:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          - CANCELLED
        requestedAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
    PublishView:
      type: object
      properties:
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
        publicUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
        requests:
          type: array
          items:
            $ref: "#/components/schemas/PublishRequestView"
        myPublicUrl:
          type: string
        requireOwnerApproval:
          type: boolean
        personalPageUrls:
          type: array
          items:
            type: string
...
    RequestPublishBody:
      type: object
      properties:
        note:
          type: string
        theme:
          type: string
    AddMemberRequest:
      required:
      - email
      type: object
      properties:
        email:
          minLength: 1
          type: string
    CreateRequest:
      required:
      - description
      type: object
      properties:
        checklistItemId:
          type: string
        category:
          type: string
          enum:
          - TRANSPORTATION
          - LODGING
          - ACTIVITIES
          - SHOPPING
          - FOOD
          - OTHERS
        description:
          maxLength: 300
          minLength: 0
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        allDay:
          type: boolean
        cost:
          type: number
        currency:
          type: string
        costCharged:
          type: boolean
        costSharedByUserIds:
          type: array
          items:
            type: string
        costPaidByUserId:
          type: string
        countryCodes:
          type: array
          items:
            type: string
        travellerIds:
          type: array
          items:
            type: string
        inheritTravellers:
          type: boolean
        note:
          type: string
        flight:
          $ref: "#/components/schemas/FlightSnapshot"
...
    PatchRequest:
      type: object
      properties:
        category:
          type: string
          enum:
          - TRANSPORTATION
          - LODGING
          - ACTIVITIES
          - SHOPPING
          - FOOD
          - OTHERS
        description:
          maxLength: 300
          minLength: 0
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        allDay:
          type: boolean
        cost:
          type: number
        currency:
          type: string
        costCharged:
          type: boolean
        costSharedByUserIds:
          type: array
          items:
            type: string
        costPaidByUserId:
          type: string
        countryCodes:
          type: array
          items:
            type: string
        travellerIds:
          type: array
          items:
            type: string
        inheritTravellers:
          type: boolean
        note:
          type: string
        flight:
          $ref: "#/components/schemas/FlightSnapshot"
...
    Choice:
      type: object
      properties:
        departureTime:
          type: string
        arrivalDate:
          type: string
        arrivalTime:
          type: string
        flight:
          $ref: "#/components/schemas/FlightSnapshot"
    LookupResponse:
      type: object
      properties:
        status:
          type: string
        operatingNumber:
          type: string
        departureTime:
          type: string
        arrivalDate:
          type: string
        arrivalTime:
          type: string
        flight:
          $ref: "#/components/schemas/FlightSnapshot"
        choices:
          type: array
          items:
            $ref: "#/components/schemas/Choice"
...
    LiveLeg:
      type: object
      properties:
        gate:
          type: string
        baggageBelt:
          type: string
        delayMinutes:
          type: integer
          format: int32
        actualLocal:
          type: string
    PublicAirport:
      type: object
      properties:
        iata:
          type: string
        name:
          type: string
        city:
          type: string
        countryCode:
          type: string
        timezone:
          type: string
    PublicFlightView:
      type: object
      properties:
        number:
          type: string
        operatingNumber:
          type: string
        schedule:
          $ref: "#/components/schemas/PublicSchedule"
        live:
          $ref: "#/components/schemas/PublicLive"
        scheduleFetchedAt:
          type: string
          format: date-time
        liveFetchedAt:
          type: string
          format: date-time
        stale:
          type: boolean
        ttlSeconds:
          type: integer
          format: int64
    PublicLive:
      type: object
      properties:
        status:
          type: string
        departure:
          $ref: "#/components/schemas/LiveLeg"
        arrival:
          $ref: "#/components/schemas/LiveLeg"
    PublicSchedule:
      type: object
      properties:
        status:
          type: string
        aircraftModel:
          type: string
        airline:
          type: string
        departure:
          $ref: "#/components/schemas/ScheduleLeg"
        arrival:
          $ref: "#/components/schemas/ScheduleLeg"
    ScheduleLeg:
      type: object
      properties:
        airport:
          $ref: "#/components/schemas/PublicAirport"
        scheduledLocal:
          type: string
        scheduledUtc:
          type: string
        revisedLocal:
          type: string
        predictedLocal:
          type: string
        terminal:
          type: string
...